+++ R/deprecated.R
@@ -3,11 +3,11 @@
 #' @inheritParams plot_strata_predictions
 #' @export
 plot_maihda_predictions_by <- function(
-    x,
-    by = NULL,
-    scale = c("response", "link"),
-    which = c("null", "adjusted"),
-    sort = TRUE
+  x,
+  by = NULL,
+  scale = c("response", "link"),
+  which = c("null", "adjusted"),
+  sort = TRUE
 ) {
   lifecycle::deprecate_warn(
     "0.12.0",