+++ NEWS.md
@@ -4,7 +4,7 @@
 
 * `plot_strata_predictions()` to plot the predictions of a MAIHDA model 
   by strata
-* new `stratified_by` argument for `plot_categorical()`
+* new `stratified_by` argument for `plot_categorical()` and `plot_continuous()`
 
 **Renamed function**