+++ NEWS.md
@@ -6,7 +6,7 @@
   by strata
 * new `bootstrap_vpc` and `bootstrap_pcv` arguments in `tbl_maihda()` and
   `tbl_partially_adjusted_maihda()`
-* `calculate_partially_adjusted_maihda()` to pre-compute partially adjusted 
+* `calculate_partially_adjusted_maihda()` to compute partially adjusted 
   MAIHDA
 * new `stratified_by` argument for `plot_categorical()`, `plot_continuous()`
   and `plot_means()`